+++ source/ts/is-not-or.ts
@@ -1,22 +1,19 @@
 //Is object exist?
-function is(obj) {
+function is(obj:any) {
   if (obj === null || obj === undefined) return false;
   if (isNaN(obj) && typeof(obj) === "number") return false;
   else return true;
 }
 
 //Not exist?
-function not(obj) {
+function not(obj:any) {
   return !is(obj);
 }
 
 //Choose first existing object
-function or(array) {
-  var k;
-  for(k in array) {
-    if (is(array[k])) {
-      return array[k];
-    }
+function or(list:Array<any>) {
+  for(let value of list) {
+    if (is(value)) return value;
   }
   return null;
 }