+++ source/ts/randtext.ts
@@ -1,5 +1,5 @@
 //Print random string
-function randtext(len) {
+function randtext(len:number):string {
   var trace:string = "";
   for(let i=0; i<len; i++) {
     trace += getrandalf();
@@ -7,8 +7,8 @@ function randtext(len) {
   return trace;
 }
 
-function getrandalf() {
-  var a = rand(0,25);
+function getrandalf():string {
+  var a:number = rand(0,25);
   
   var alf = ['a','b','c','d','e','f','g','h','i','j','k','l','m','n','o','p','q','r','s','t','u','v','w','x','y','z'];